Decouple OCR service from input helpers
